requirements.txt
streamlit==0.88.0
requests==2.26.0
animals.py
dadjokes==1.3.2
joke_generator==1.0.1
randfacts==0.10.0
streamlit_lottie==0.0.2
streamlit_observable==0.0.8
